test: harden production probe policy defaults
